Gas Sensor & Temperature sensor
since we working on this project to explore Mars. So  here Gas sensors and Temperature Sensors play a major  role in the exploration
as Gas sensor helps in finding different types and concentrations of gas present on Mars.
A temperature sensor helps to determine the temperature on Mars that is how hot the mars is.